Many tweaks to build under Windows.  We still have a long way to go before this works.  In particular, we have many dependencies on the argp command line argument processing support from glibc.  I've partially stubbed it out for Windows builds, but that is not sufficient because that ends up omitting some important global variable initializations.

+++ src/utils.cc
@@ -13,14 +13,14 @@ using std::string;
 pred_stat compute_pred_stat(int s, int f, int os, int of, int confidence)
 {
     pred_stat p;
-    p.fs = ((float)  f) / ( s +  f);
-    p.co = ((float) of) / (os + of);
+    p.fs = ((double)  f) / ( s +  f);
+    p.co = ((double) of) / (os + of);
     p.in = p.fs - p.co;
     p.lb = p.in - Confidence::critical(confidence) * sqrt(((p.fs * (1 - p.fs)) / (f + s)) + ((p.co * (1 - p.co))/(of + os)));
     return p;
 }
 
-bool retain_pred(int s, int f, float lb) { return lb > 0 && s + f >= 10; }
+bool retain_pred(int s, int f, double lb) { return lb > 0 && s + f >= 10; }
 
 bool retain_pred(int s, int f, int os, int of, int confidence)
 {
